[web]Don't expose Firebase username and password to the client
I will send a follow-on PR to replace the secrets.
Tested locally with `packages/web $ yarn run dev`
